% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciinema.R
\name{asciicast_base64}
\alias{asciicast_base64}
\title{encode a -n asciicast json file in base64}
\usage{
asciicast_base64(file)
}
\arguments{
\item{file}{asciicast file}
}
\value{
the file encoded in base64 with content type `application/json`
}
\description{
encode a -n asciicast json file in base64
}
