% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast_hook.R
\name{asciicast_hook}
\alias{asciicast_hook}
\title{asciicast hook}
\usage{
asciicast_hook(options)
}
\arguments{
\item{options}{chunk options}
}
\description{
Use this option hook to produce an asciicast
with the R code in the chunk
}
