% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciinemaPlayer.R
\name{asciinemaPlayer}
\alias{asciinemaPlayer}
\title{player for asciicasts}
\usage{
asciinemaPlayer(file, src = asciicast_base64(file), width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{file}{asciicast json file}

\item{src}{source asciicast}

\item{width}{width}

\item{height}{height}

\item{elementId}{id}
}
\description{
player for asciicasts
}
\examples{
\dontrun{
  asciinemaPlayer( system.file("resources", "mapscii.json", package = "rsciinema") )
}

}
