% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backspace.R
\name{backspace}
\alias{backspace}
\title{Backspace character for Casting}
\usage{
backspace(n = 1)
}
\arguments{
\item{n}{Number of backspaces to put in}
}
\value{
A character vector
}
\description{
Backspace character for Casting
}
\examples{
backspace(4)
}
