% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R
\name{json_asciicast}
\alias{json_asciicast}
\title{convert an asciicast tibble to its json representation}
\usage{
json_asciicast(data)
}
\arguments{
\item{data}{an asciicast tibble}
}
\value{
json formatted asciicast
}
\description{
convert an asciicast tibble to its json representation
}
