% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_themes.R
\docType{data}
\name{rstudio_themes}
\alias{rstudio_themes}
\title{RStudio Syntax Highlighting Themes}
\format{A list with 31 elements, which are each a tibble with the colmns:
\describe{
\item{rule}{CSS tag}
\item{property}{what kind of property is it}
\item{value}{color to display}
}}
\usage{
rstudio_themes
}
\description{
RStudio Syntax Highlighting Themes
}
\keyword{datasets}
