% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R
\name{write_asciicast}
\alias{write_asciicast}
\title{write ascii cast tibble to a json file}
\usage{
write_asciicast(data, path)
}
\arguments{
\item{data}{asciicast tibble}

\item{path}{output file}
}
\description{
write ascii cast tibble to a json file
}
