% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciinema.R
\name{asciinema}
\alias{asciinema}
\title{player for asciicasts}
\usage{
asciinema(file, cols = 80, rows = 24, autoplay = FALSE, loop = FALSE,
  start_at = 0, speed = 1, poster_text = NULL, poster_frame = NULL,
  font_size = "small", src = asciicast_base64(file), width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{file}{asciicast json file}

\item{cols}{number of columns of players terminal}

\item{rows}{number of rows of players terminal}

\item{autoplay}{if `TRUE` it autoplays}

\item{loop}{if `TRUE` it loops}

\item{start_at}{a number of seconds or a `Period` created by e.g. [lubridate::seconds()]}

\item{speed}{speed, 2 means twice as fast}

\item{poster_text}{if not `NULL`, used as the text of the poster (preview)}

\item{poster_frame}{if not `NULL`, used as the}

\item{font_size}{size of terminal font. Possible values: small, medium, big, any css `font-size` value (e.g. 15px)}

\item{src}{source asciicast}

\item{width}{width}

\item{height}{height}

\item{elementId}{id}
}
\description{
player for asciicasts
}
\examples{
\dontrun{
  asciinema( system.file("resources", "mapscii.json", package = "rsciinema") )
}

}
