% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciinema_document.R
\name{asciinema_document}
\alias{asciinema_document}
\title{Format adapter to embed asciicast into html documents}
\usage{
asciinema_document(format = rmarkdown::html_document, width = 80,
  height = 24, speed = 0.1, ...)
}
\arguments{
\item{format}{base format, e.g. [rmarkdown::html_document()]}

\item{width}{Number of columns of player's terminal}

\item{height}{Number of lines of player's terminal}

\item{speed}{Typing speed in seconds. The average number of seconds it takes to type one character}

\item{...}{passed to the base format}
}
\description{
Format adapter to embed asciicast into html documents
}
