% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciinemaPlayer.R
\name{asciinemaPlayer}
\alias{asciinemaPlayer}
\title{player for asciicasts}
\usage{
asciinemaPlayer(file, cols = 80, rows = 24, autoplay = FALSE,
  loop = FALSE, start_at = 0, speed = 1, src = asciicast_base64(file),
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{file}{asciicast json file}

\item{cols}{number of columns of players terminal}

\item{rows}{number of rows of players terminal}

\item{autoplay}{if `TRUE` it autoplays}

\item{loop}{if `TRUE` it loops}

\item{start_at}{a number of seconds or a `Period` created by e.g. [lubridate::seconds()]}

\item{speed}{speed, 2 means twice as fast}

\item{src}{source asciicast}

\item{width}{width}

\item{height}{height}

\item{elementId}{id}
}
\description{
player for asciicasts
}
\examples{
\dontrun{
  asciinemaPlayer( system.file("resources", "mapscii.json", package = "rsciinema") )
}

}
